Move 'get_config' to config module
